MAIN - arch_sci_type uint16

+++ xen/include/public/arch-arm.h
@@ -319,7 +319,7 @@ struct xen_arch_domainconfig {
     /* IN */
     uint16_t tee_type;
     /* IN */
-    uint8_t arm_sci_type;
+    uint16_t arm_sci_type;
     /* IN */
     uint32_t nr_spis;
     /*
